add brew screen with temperature updates add new PID routine update PWM actuator to optimize dury time various bugfixes

+++ oinkbrew/Configuration.h
@@ -67,7 +67,7 @@ public:
     static bool loadEguiSettings();
     static void storeEguiSettings();
 
-    static void clear(uint8_t* p, uint8_t size);
+    static void clear(uint8_t* p, uint16_t size);
 };
 
 extern Configuration conf;
